requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pytest==8.2.2
httpx==0.27.0
python-dotenv==1.0.1
